task #119 GH#19 S3: feed %zu the size_t it names, not an unsigned long
The Windows gate compiled tests/host/test_gh19_s3_pulse_product_acceptance.cpp for the
first time and stopped on C2220 (C4477 promoted by /WX): the sync summary printf passes two
`%zu` conversions but both arguments were cast to `unsigned long`. That cast is a no-op on
LP64 (Linux/macOS, where unsigned long is 64-bit) and a genuine type mismatch on LLP64
(Windows, where unsigned long is 32-bit while size_t is 64-bit), so no local clang or gcc
build could see it. Both locals are already `std::size_t` -- which is exactly what `%zu`
names -- so the cast is deleted, matching the two other %zu printf sites in this same file
(lines 855 and 904) that already pass size_t directly.

The file is byte-identical from the frozen parent 0bfbf87 through c3427e6, and 0bfbf87
never ran CI, so this is a latent defect in task #119's own acceptance test that the remote
MSVC job surfaced now -- not a regression introduced by the delivery commit.

Output-neutral by construction, and verified to be: the Release suite still reports
"[test_gh19_s3_pulse_product_acceptance] 122 checks OK" at rc 0, and the regenerated run is
byte-identical to the committed report/gh19-s3-pulse-product/product_acceptance_run.txt, so
no evidence artifact moves. The 10-mutant matrix mutates this file, so it was re-run to
confirm every anchor still applies.

c3427e6 and 0bfbf87 remain as published; this commit stacks on top of them.

+++ tests/host/test_gh19_s3_pulse_product_acceptance.cpp
@@ -1263,8 +1263,8 @@ void hard_sync_reset_is_sized_on_the_emitted_signal() {
     std::printf(
         "  sync: hysteresis reference rise at frame %zu (thr=%.4f hyst=%.4f, holds above %.4f; "
         "frame 0 primes) | first output divergence at frame %zu | sync=%.12f  0.5*free=%.12f\n",
-        static_cast<unsigned long>(predictedEdgeFrame), syncThr, syncHyst, syncThr - syncHyst,
-        static_cast<unsigned long>(firstDivergence), aSync[predictedEdgeFrame],
+        predictedEdgeFrame, syncThr, syncHyst, syncThr - syncHyst,
+        firstDivergence, aSync[predictedEdgeFrame],
         0.5 * aFree[predictedEdgeFrame]);
     // The reset sample is `emittedAt(0) - 0.5*(emittedAt(0) - free[r])`, and at a 50% duty with
     // w < 0.5 the two-edge formula makes emittedAt(0) exactly 0 (naive +1 cancels the residual of
